+++ src/test/java/uk/co/reecedunn/intellij/plugin/xquery/tests/parser/PluginParserTest.kt
@@ -2216,4 +2216,36 @@ private class PluginParserTest : ParserTestCase() {
             }
         }
     }
+
+    @Nested
+    @DisplayName("XQuery IntelliJ Plugin EBNF (93) ElvisExpr")
+    internal inner class ElvisExpr {
+        @Test
+        @DisplayName("elvis")
+        fun elvis() {
+            val expected = loadResource("tests/parser/xpath-ng/proposal-2/ElvisExpr.txt")
+            val actual = parseResource("tests/parser/xpath-ng/proposal-2/ElvisExpr.xq")
+            assertThat(prettyPrintASTNode(actual), `is`(expected))
+        }
+
+        @Test
+        @DisplayName("elvis; compact whitespace")
+        fun elvis_CompactWhitespace() {
+            val expected = loadResource("tests/parser/xpath-ng/proposal-2/ElvisExpr_CompactWhitespace.txt")
+            val actual = parseResource("tests/parser/xpath-ng/proposal-2/ElvisExpr_CompactWhitespace.xq")
+            assertThat(prettyPrintASTNode(actual), `is`(expected))
+        }
+
+        @Nested
+        @DisplayName("error recovery; missing token")
+        internal inner class MissingToken {
+            @Test
+            @DisplayName("missing else Expr")
+            fun missingElseExpr() {
+                val expected = loadResource("tests/parser/xpath-ng/proposal-2/ElvisExpr_MissingElseExpr.txt")
+                val actual = parseResource("tests/parser/xpath-ng/proposal-2/ElvisExpr_MissingElseExpr.xq")
+                assertThat(prettyPrintASTNode(actual), `is`(expected))
+            }
+        }
+    }
 }